ST-4756 migration and automagic to find .yml
