+++ src/test/resources/npl-sources/success/multiple_files/src/main/npl/processes/settle.npl
@@ -5,8 +5,8 @@ use objects.iou.Iou
 @api
 protocol[iouOwner, carOwner] Settle(var iou: Iou) {
     @api
-    permission[iouOwner | carOwner] swap() {
+    permission[iouOwner | carOwner] swap() returns List<Number> {
         // Transfer logic.
-        iou.forgive[iouOwner]();
+        return listOf(1,2,3,4);
     };
 };
